Split BlueapiClient, Devices, and plans
